Apply API design review fixes to com.landawn.abacus.cache
  Implements the actionable findings from the package API review, staged
  P1→P3. All 342 tests pass (0 failures/errors).

  P1 — consistency wins:
  - CaffeineCache.stats() now returns the framework CacheStats
(consistent
    with LocalCache); add caffeineStats() for the native Caffeine
object.
  - Document Ehcache's intentional lack of stats() (Ehcache 3.x exposes
    statistics only via a StatisticsService, not the wrapped Cache).
  - Drop misused @Deprecated on the always-unsupported
CaffeineCache.keySet()
    and Ehcache.keySet()/size() (keep the
UnsupportedOperationException).
  - Ehcache.putAll validates via N.checkElementNotNull like
getAll/removeAll.
  - Add zero-guarded hitRate()/missRate() to CacheStats and
OffHeapCacheStats.
  - KryoTranscoder: add KryoParser-injecting constructors (enables class
    pre-registration); parser is now an instance field.
  - OffHeapStore extends AutoCloseable with a default close(), invoked
from
    AbstractOffHeapCache.close() to release store resources.

  P2 — design changes:
  - Widen incr/decr delta int -> long across DistributedCacheClient,
    AbstractJedisCacheClient and SpyMemcached (backends are 64-bit).
  - CacheFactory: add createOffHeapCache (x3), createCaffeineCache,
    createEhcache; reorder the pool createLocalCache overload to
pool-first.
  - Rename OffHeapCache25 -> ForeignMemoryOffHeapCache (capability-based
name,
    no JDK-version leak).
  - ChronicleMap: @Deprecated(since="2.8.1", forRemoval=true).

  P3 — cleanups:
  - LocalCache now rejects null values (aligns with
Caffeine/Ehcache/OffHeap).
  - OffHeapCacheStats: rename hitCountByDisk -> hitCountFromDisk; remove
the
    unused public OccupiedSlot record.
  - AbstractOffHeapCache: evict() -> private reclaimEmptySegments();
reorder
    copyToMemory params to (startPtr, srcBytes, srcOffset, len) to
mirror
    copyFromMemory.

  Tests updated for the breaking changes accordingly.

  BREAKING CHANGES:
  - DistributedCacheClient.incr/decr(String,int) -> (String,long).
  - OffHeapCache25 renamed to ForeignMemoryOffHeapCache.
  - OffHeapCacheStats component hitCountByDisk renamed to
hitCountFromDisk;
    nested OccupiedSlot record removed.
  - CaffeineCache.stats() return type changed to
com.landawn.abacus.cache.CacheStats
    (native object now via caffeineStats()).
  - CacheFactory.createLocalCache(long,long,KeyedObjectPool) reordered
to
    createLocalCache(KeyedObjectPool,long,long).
  - LocalCache.put now rejects null values with
IllegalArgumentException.

+++ src/main/java/com/landawn/abacus/cache/AbstractJedisCacheClient.java
@@ -332,7 +332,7 @@ abstract class AbstractJedisCacheClient<T> extends AbstractDistributedCacheClien
      * @return the value after increment (will be 1 if the key did not exist before)
      * @throws IllegalArgumentException if {@code key} is {@code null}
      * @throws RuntimeException if a network error, timeout occurs, or if the key contains a non-integer value
-     * @see #incr(String, int)
+     * @see #incr(String, long)
      * @see #decr(String)
      */
     @Override
@@ -369,10 +369,10 @@ abstract class AbstractJedisCacheClient<T> extends AbstractDistributedCacheClien
      * @throws IllegalArgumentException if {@code key} is {@code null} or {@code delta} is negative
      * @throws RuntimeException if a network error, timeout occurs, or if the key contains a non-integer value
      * @see #incr(String)
-     * @see #decr(String, int)
+     * @see #decr(String, long)
      */
     @Override
-    public long incr(final String key, final int delta) {
+    public long incr(final String key, final long delta) {
         N.checkArgNotNegative(delta, "delta");
 
         final byte[] keyBytes = getKeyBytes(key);
@@ -403,7 +403,7 @@ abstract class AbstractJedisCacheClient<T> extends AbstractDistributedCacheClien
      * @return the value after decrement (can be negative in Redis, will be -1 if the key did not exist before)
      * @throws IllegalArgumentException if {@code key} is {@code null}
      * @throws RuntimeException if a network error, timeout occurs, or if the key contains a non-integer value
-     * @see #decr(String, int)
+     * @see #decr(String, long)
      * @see #incr(String)
      */
     @Override
@@ -442,10 +442,10 @@ abstract class AbstractJedisCacheClient<T> extends AbstractDistributedCacheClien
      * @throws IllegalArgumentException if {@code key} is {@code null} or {@code delta} is negative
      * @throws RuntimeException if a network error, timeout occurs, or if the key contains a non-integer value
      * @see #decr(String)
-     * @see #incr(String, int)
+     * @see #incr(String, long)
      */
     @Override
-    public long decr(final String key, final int delta) {
+    public long decr(final String key, final long delta) {
         N.checkArgNotNegative(delta, "delta");
 
         final byte[] keyBytes = getKeyBytes(key);
